Fix error handling in Start-Process function by ensuring temporary files are removed only if they exist

+++ scripts/powershell/deploy-adf-integration-runtime.ps1
@@ -70,8 +70,13 @@ function Start-Process([string] $process, [string] $arguments)
 	$errContent = [string] (Get-Content -Path $errorFile -Delimiter "!!!DoesNotExist!!!")
 	$outContent = [string] (Get-Content -Path $outFile -Delimiter "!!!DoesNotExist!!!")
 
+	if (Test-Path $errorFile) {
 	Remove-Item $errorFile
-	Remove-Item $outFile
+	}
+
+	if (Test-Path $outFile) {
+	Remove-Item $outFile 
+	}
 
 	if($proc.ExitCode -ne 0 -or $errVariable -ne "")
 	{
